Moved ChargeCodeActivity enum to separate file
