+++ src/angular/hq/src/app/enums/charge-code-activity.ts
@@ -1,0 +1,6 @@
+export enum ChargeCodeActivity {
+  General = 1,
+  Project = 2,
+  Quote = 3,
+  Service = 4,
+}
